requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
cx_Oracle==8.3.0
Pillow==8.3.1
opencv-python==4.5.3.56
numpy==1.21.2
cryptography==41.0.5
PyQt5==5.15.4
cheroot==8.6.0